STY: Moved test
