+++ OMMBV/tests/test_former_methods.py
@@ -1358,3 +1358,241 @@ class TestIntegratedMethods():
 
         except:
             pass
+
+    def test_heritage_geomag_efield_scalars_plots(self):
+        """Summary plots of the heritage code path for scaling electric fields and ion drifts"""
+        import matplotlib.pyplot as plt
+
+        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
+        # data returned are the locations along each direction
+        # the full range of points obtained by iterating over all
+        # recasting alts into a more convenient form for later calculation
+        p_alts = [p_alts[0]]*len(p_longs)
+
+        north_zonal = np.zeros((len(p_lats), len(p_longs) + 1))
+        north_mer = north_zonal.copy()
+        south_zonal = north_zonal.copy()
+        south_mer = north_zonal.copy()
+        eq_zonal = north_zonal.copy()
+        eq_mer = north_zonal.copy()
+
+        north_zonald = np.zeros((len(p_lats), len(p_longs) + 1))
+        north_merd = north_zonal.copy()
+        south_zonald = north_zonal.copy()
+        south_merd = north_zonal.copy()
+        eq_zonald = north_zonal.copy()
+        eq_merd = north_zonal.copy()
+
+        date = dt.datetime(2000, 1, 1)
+        # set up multi
+        if self.dc is not None:
+            targets = itertools.cycle(dc.ids)
+            pending = []
+            for i, p_lat in enumerate(p_lats):
+                # iterate through target cyclicly and run commands
+                print (i, p_lat)
+                dview.targets = next(targets)
+                pending.append(
+                    dview.apply_async(OMMBV.heritage_scalars_for_mapping_ion_drifts,
+                                      [p_lat]*len(p_longs), p_longs,
+                                      p_alts, [date]*len(p_longs)))
+            for i, p_lat in enumerate(p_lats):
+                print ('collecting ', i, p_lat)
+                # collect output
+                scalars = pending.pop(0).get()
+                north_zonal[i, :-1] = scalars['north_mer_fields_scalar']
+                north_mer[i, :-1] = scalars['north_zon_fields_scalar']
+                south_zonal[i, :-1] = scalars['south_mer_fields_scalar']
+                south_mer[i, :-1] = scalars['south_zon_fields_scalar']
+                eq_zonal[i, :-1] = scalars['equator_mer_fields_scalar']
+                eq_mer[i, :-1] = scalars['equator_zon_fields_scalar']
+
+                north_zonald[i, :-1] = scalars['north_zonal_drifts_scalar']
+                north_merd[i, :-1] = scalars['north_mer_drifts_scalar']
+                south_zonald[i, :-1] = scalars['south_zonal_drifts_scalar']
+                south_merd[i, :-1] = scalars['south_mer_drifts_scalar']
+                eq_zonald[i, :-1] = scalars['equator_zonal_drifts_scalar']
+                eq_merd[i, :-1] = scalars['equator_mer_drifts_scalar']
+        else:
+            for i, p_lat in enumerate(p_lats):
+                print (i, p_lat)
+                scalars = OMMBV.heritage_scalars_for_mapping_ion_drifts([p_lat]*len(p_longs), p_longs,
+                                                                       p_alts, [date]*len(p_longs))
+                north_zonal[i, :-1] = scalars['north_mer_fields_scalar']
+                north_mer[i, :-1] = scalars['north_zon_fields_scalar']
+                south_zonal[i, :-1] = scalars['south_mer_fields_scalar']
+                south_mer[i, :-1] = scalars['south_zon_fields_scalar']
+                eq_zonal[i, :-1] = scalars['equator_mer_fields_scalar']
+                eq_mer[i, :-1] = scalars['equator_zon_fields_scalar']
+
+                north_zonald[i, :-1] = scalars['north_zonal_drifts_scalar']
+                north_merd[i, :-1] = scalars['north_mer_drifts_scalar']
+                south_zonald[i, :-1] = scalars['south_zonal_drifts_scalar']
+                south_merd[i, :-1] = scalars['south_mer_drifts_scalar']
+                eq_zonald[i, :-1] = scalars['equator_zonal_drifts_scalar']
+                eq_merd[i, :-1] = scalars['equator_mer_drifts_scalar']
+        # account for periodicity
+        north_zonal[:, -1] = north_zonal[:, 0]
+        north_mer[:, -1] = north_mer[:, 0]
+        south_zonal[:, -1] = south_zonal[:, 0]
+        south_mer[:, -1] = south_mer[:, 0]
+        eq_zonal[:, -1] = eq_zonal[:, 0]
+        eq_mer[:, -1] = eq_mer[:, 0]
+        north_zonald[:, -1] = north_zonald[:, 0]
+        north_merd[:, -1] = north_merd[:, 0]
+        south_zonald[:, -1] = south_zonald[:, 0]
+        south_merd[:, -1] = south_merd[:, 0]
+        eq_zonald[:, -1] = eq_zonald[:, 0]
+        eq_merd[:, -1] = eq_merd[:, 0]
+
+        xtickvals = ['-25', '-12.5', '0', '12.5', '25']
+        xtickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats) - 1)
+        ytickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
+
+        try:
+            fig = plt.figure()
+            plt.imshow(eq_zonal, origin='lower')  # , vmin=0, vmax=2)
+            plt.colorbar()
+            plt.yticks(xtickarr, xtickvals)
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Meridional Electric Field Mapping to Magnetic Equator')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('eq_mer_field_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(eq_mer, origin='lower')  # , vmin=0, vmax=1.)
+            plt.colorbar()
+            plt.yticks(xtickarr, ['-50', '-25', '0', '25', '50'])
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Zonal Electric Field Mapping to Magnetic Equator')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('eq_zon_field_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(north_zonal, origin='lower')  # , vmin=0, vmax=2)
+            plt.colorbar()
+            plt.yticks(xtickarr, ['-50', '-25', '0', '25', '50'])
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Meridional Electric Field Mapping to Northern Footpoint')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('north_mer_field_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(north_mer, origin='lower')  # , vmin=0, vmax=2)
+            plt.colorbar()
+            plt.yticks(xtickarr, ['-50', '-25', '0', '25', '50'])
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Zonal Electric Field Mapping to Northern Footpoint')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('north_zon_field_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(south_zonal, origin='lower')  # , vmin=0, vmax=2)
+            plt.colorbar()
+            plt.yticks(xtickarr, ['-50', '-25', '0', '25', '50'])
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Meridional Electric Field Mapping to Southern Footpoint')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('south_mer_field_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(south_mer, origin='lower')  # , vmin=0, vmax=2)
+            plt.colorbar()
+            plt.yticks(xtickarr, ['-50', '-25', '0', '25', '50'])
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Zonal Electric Field Mapping to Southern Footpoint')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('south_zon_field_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(np.log10(eq_zonald), origin='lower')  # , vmin=0, vmax=2)
+            plt.colorbar()
+            plt.yticks(xtickarr, xtickvals)
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Log Zonal Ion Drift Mapping to Magnetic Equator')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('eq_zonal_drift_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(np.log10(eq_merd), origin='lower')  # , vmin=0, vmax=1.)
+            plt.colorbar()
+            plt.yticks(xtickarr, xtickvals)
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Log Meridional Ion Drift Mapping to Magnetic Equator')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('eq_mer_drift_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(north_zonald, origin='lower')  # , vmin=0, vmax=2)
+            plt.colorbar()
+            plt.yticks(xtickarr, xtickvals)
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Zonal Ion Drift Mapping to Northern Footpoint')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('north_zonal_drift_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(north_merd, origin='lower')  # , vmin=0, vmax=2)
+            plt.colorbar()
+            plt.yticks(xtickarr, xtickvals)
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Meridional Ion Drift Mapping to Northern Footpoint')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('north_mer_drift_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(south_zonald, origin='lower')  # , vmin=0, vmax=2)
+            plt.colorbar()
+            plt.yticks(xtickarr, xtickvals)
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Zonal Ion Drift Mapping to Southern Footpoint')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('south_zonal_drift_heritage.pdf')
+            plt.close()
+
+            fig = plt.figure()
+            plt.imshow(south_merd, origin='lower')  # , vmin=0, vmax=2)
+            plt.colorbar()
+            plt.yticks(xtickarr, xtickvals)
+            plt.xticks(ytickarr, ['0', '72', '144', '216', '288', '360'])
+            plt.title('Meridional Ion Drift Mapping to Southern Footpoint')
+            plt.xlabel('Geodetic Longitude (Degrees)')
+            plt.ylabel('Geodetic Latitude (Degrees)')
+            plt.tight_layout()
+            plt.savefig('south_mer_drift_heritage.pdf')
+            plt.close()
+
+        except:
+            pass
